requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
contourpy==1.1.1
cycler==0.11.0
defusedxml==0.7.1
fonttools==4.42.1
idna==3.4
kiwisolver==1.4.5
markdown-it-py==3.0.0
matplotlib==3.8.0
mdurl==0.1.2
networkx==3.1
numpy==1.26.0
odfpy==1.4.1
packaging==23.1
pandas==2.1.0
Pillow==10.0.1
Pygments==2.16.1
pygraphviz==1.11
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
rich==13.5.3
six==1.16.0
tzdata==2023.3
urllib3==2.0.4
